+++ Cargo.toml
@@ -8,6 +8,7 @@ members = [
     "client",
     "core",
     "drone",
+    "perf",
     "validator",
     "genesis",
     "genesis_programs",
