+++ core/Cargo.toml
@@ -57,6 +57,7 @@ solana-merkle-tree = { path = "../merkle-tree", version = "0.21.0" }
 solana-metrics = { path = "../metrics", version = "0.21.0" }
 solana-measure = { path = "../measure", version = "0.21.0" }
 solana-netutil = { path = "../netutil", version = "0.21.0" }
+solana-perf = { path = "../perf", version = "0.21.0" }
 solana-runtime = { path = "../runtime", version = "0.21.0" }
 solana-sdk = { path = "../sdk", version = "0.21.0" }
 solana-stake-api = { path = "../programs/stake_api", version = "0.21.0" }
