+++ core/src/shred_fetch_stage.rs
@@ -1,10 +1,10 @@
 //! The `shred_fetch_stage` pulls shreds from UDP sockets and sends it to a channel.
 
-use crate::cuda_runtime::PinnedVec;
-use crate::packet::Packet;
-use crate::recycler::Recycler;
 use crate::service::Service;
 use crate::streamer::{self, PacketReceiver, PacketSender};
+use solana_ledger::packet::Packet;
+use solana_perf::cuda_runtime::PinnedVec;
+use solana_perf::recycler::Recycler;
 use std::net::UdpSocket;
 use std::sync::atomic::AtomicBool;
 use std::sync::mpsc::channel;
