+++ core/src/sigverify.rs
@@ -4,14 +4,14 @@
 //! to the GPU.
 //!
 
-use crate::cuda_runtime::PinnedVec;
 use crate::packet::{Packet, Packets};
-use crate::recycler::Recycler;
 use crate::sigverify_stage::SigVerifier;
 use bincode::serialized_size;
 use rayon::ThreadPool;
-use solana_ledger::perf_libs;
 use solana_metrics::inc_new_counter_debug;
+use solana_perf::cuda_runtime::PinnedVec;
+use solana_perf::perf_libs;
+use solana_perf::recycler::Recycler;
 use solana_sdk::message::MessageHeader;
 use solana_sdk::pubkey::Pubkey;
 use solana_sdk::short_vec::decode_len;
@@ -421,9 +421,8 @@ pub fn make_packet_from_transaction(tx: Transaction) -> Packet {
 
 #[cfg(test)]
 mod tests {
-    use super::PacketError;
+    use super::*;
     use crate::packet::{Packet, Packets};
-    use crate::recycler::Recycler;
     use crate::sigverify;
     use crate::sigverify::PacketOffsets;
     use crate::test_tx::{test_multisig_tx, test_tx};
