+++ core/src/validator.rs
@@ -128,7 +128,7 @@ impl Validator {
         warn!("vote pubkey: {:?}", vote_account);
         warn!(
             "CUDA is {}abled",
-            if solana_ledger::perf_libs::api().is_some() {
+            if solana_perf::perf_libs::api().is_some() {
                 "en"
             } else {
                 "dis"
