+++ ledger/src/entry.rs
@@ -2,7 +2,6 @@
 //! unique ID that is the hash of the Entry before it, plus the hash of the
 //! transactions within it. Entries cannot be reordered, and its field `num_hashes`
 //! represents an approximate amount of time since the last Entry was created.
-use crate::perf_libs;
 use crate::poh::Poh;
 use log::*;
 use rayon::prelude::*;
@@ -11,6 +10,7 @@ use serde::{Deserialize, Serialize};
 use solana_measure::measure::Measure;
 use solana_merkle_tree::MerkleTree;
 use solana_metrics::*;
+use solana_perf::perf_libs;
 use solana_rayon_threadlimit::get_thread_count;
 use solana_sdk::hash::Hash;
 use solana_sdk::timing;
