+++ perf/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "solana-perf"
+version = "0.21.0"
+description = "Solana Performance APIs"
+authors = ["Solana Maintainers <maintainers@solana.com>"]
+repository = "https://github.com/solana-labs/solana"
+license = "Apache-2.0"
+homepage = "https://solana.com/"
+edition = "2018"
+
+[dependencies]
+rand = "0.6.5"
+dlopen = "0.1.8"
+dlopen_derive = "0.1.4"
+log = "0.4.8"
+solana-sdk = { path = "../sdk", version = "0.21.0" }
+
+[lib]
+name = "solana_perf"
+
