+++ perf/src/lib.rs
@@ -1,0 +1,6 @@
+pub mod cuda_runtime;
+pub mod perf_libs;
+pub mod recycler;
+
+#[macro_use]
+extern crate log;
